% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{data_overview}
\alias{data_overview}
\title{data_overview}
\usage{
data_overview(data_set, type_infer = TRUE, out_file = "data_description")
}
\arguments{
\item{data_set}{The data set}

\item{type_infer}{\code{TRUE}(default)/\code{FALSE}. If \code{TRUE}, the type
of the column will be inferred based on the pattern of the content. Otherwise,
the type will be the storing type of the column.}

\item{out_file}{The output filename. If \code{NULL} or \code{NA} given, no
output file will be generated, and the result will be returned as a
\code{data.frame}, used for debugging.}
}
\value{
res If \code{out_file} is \code{NULL} or \code{NA}, a data.frame will
be returned. Otherwise, nothing returned but an excel file will be generated.
}
\description{
\code{data_overview} generates an excel file describe the data's
basic information like the dimensions, the type of each column, basic stats
and so on. Most importantly, the following precedure will base on the
(revised) excel file to perform data cleansing.
}
