% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{logistic_model}
\alias{logistic_model}
\title{Stepwise Logistic Regression}
\usage{
logistic_model(df, y, summary_woe, max_nvars = NULL, parallel = FALSE, ...)
}
\description{
Stepwise Logistic Regression
}
