% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProbAdjust.R
\name{PostProbAdjust}
\alias{PostProbAdjust}
\title{Posterior Probability Adjustment}
\usage{
PostProbAdjust(prob, r0 = 0.05, r1 = 0.5)
}
\arguments{
\item{prob}{prob to be bad given by model prediction}

\item{r0}{the original (designated) bad rate}

\item{r1}{the bad rate in training set (after oversampling)}
}
\description{
Adjust for posterior probability, considering the original
bad rate and the bad rate after oversampling.
}
\examples{
\dontrun{
# Given the predicted probability 0.67, the original bad rate 0.05, and the
# bad rate after oversampling 0.5, then the adjusted posterior probability
# to be bad:l
PostProbAdjust(0.67, 0.05, 0.5)
}

}
