% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(data_set, xlsx_file = "data_description",
  sheetName = "Data Overview", max_levels = 20, max_uniratio = 0.95,
  max_na_ratio = 0.5)
}
\arguments{
\item{data_set}{The dataset}

\item{xlsx_file}{The excel file generated by \code{data_overview}}

\item{sheetName}{The name of the sheet in the excel file}

\item{max_levels}{The tolerate maximum number of levels for a categorical
variable}

\item{max_uniratio}{The tolerate maximum ratio of a single value in a var}

\item{max_na_ratio}{The tolerate maximum ratio of missing value in a var}
}
\value{
data.frame the cleaned \code{data.frame}
}
\description{
\code{preprocess} is a utility function which help to generate a
dataset based on the original dataset in accordance with the description xlsx
file.
}
