% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{woe}
\alias{woe}
\alias{woe.default}
\alias{woe.data.frame}
\alias{SummaryWoe}
\alias{ReplaceWoe}
\title{Weight of Evidence}
\usage{
woe(...)

\method{woe}{default}(predictor, response, good = NA, ...)

\method{woe}{data.frame}(df, x, y, ...)

SummaryWoe(df, y, ...)

ReplaceWoe(df, y, summary_woe)
}
\arguments{
\item{predictor}{a vector representing the predictor}

\item{response}{a vecotr representing the response}

\item{good}{the value representing 'good' in \code{response}, other
values representing bad.}

\item{df}{a \code{data.frame} containing the predictor and the
response}

\item{x}{the name of the predictor in \code{df}}

\item{y}{the name of the response in \code{df}}

\item{summary_woe}{result given by SummaryWoe}
}
\value{
a \code{data.frame} showing the WoE of each level, and encapsulated
as a \code{woe} object
}
\description{
\code{woe} calculates the weight of evidence of each level in a
discrete variable with respect to a response.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default generic

\item \code{data.frame}: generic for data.frame
}}

\examples{
\dontrun{
library(ISLR)
# data.frame generic
woe(Default, 'student', 'default', good = 'No')
# default generic
woe(Default$student, Default$default, good = 'No')
SummaryWoe(Default, 'default', good = 'No')
}

}
