% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{logistic_model}
\alias{logistic_model}
\title{Stepwise Logistic Regression}
\usage{
logistic_model(df, y, summary_woe, max_nvars = NULL, seed = NULL,
  parallel = FALSE, cv_nfold = 5, auc_revise = 0, ...)
}
\description{
Stepwise Logistic Regression
}
